+++ server.js
@@ -206,6 +206,7 @@ app.use('/pl/shibcallback', require('./pages/authCallbackShib/authCallbackShib')
 app.use('/pl/azure_login', require('./pages/authLoginAzure/authLoginAzure'));
 app.use('/pl/azure_callback', require('./pages/authCallbackAzure/authCallbackAzure'));
 app.use('/pl/lti', require('./pages/authCallbackLti/authCallbackLti'));
+app.use('/pl/devlogin', require('./pages/authLoginDev/authLoginDev'));
 app.use('/pl/login', require('./pages/authLogin/authLogin'));
 // disable SEB until we can fix the mcrypt issues
 // app.use('/pl/downloadSEBConfig', require('./pages/studentSEBConfig/studentSEBConfig'));